TODO/pyproject.toml
[project]
name = "ai-stock-analyst"
version = "1.0.0"
description = "AI Stock Analyst using PydanticAI, Claude, and MCP for real-time stock analysis"
dependencies = [
    "pydantic-ai-slim[anthropic,mcp]",
    "mcp[cli]",
    "httpx",
    "pydantic",
    "logfire[cli]>=3.18.0",
    "python-dotenv>=1.1.0",
    "aiohttp>=3.12.10",
    "pandas>=2.3.0",
]
requires-python = ">=3.10"

[project.scripts]
stock-analyst = "stock_analyst.main:run"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "mypy",
]

[tool.hatch.build.targets.wheel]
packages = ["stock_analyst"] 
